+++ packages/zone.js/lib/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@npm//@bazel/concatjs:index.bzl", "ts_library")
+load("//tools:defaults.bzl", "ts_library")
 
 package(default_visibility = ["//visibility:public"])
 
@@ -27,6 +27,5 @@ ts_library(
         "@npm//@types/jasmine",
         "@npm//@types/node",
         "@npm//rxjs",
-        "@npm//tslib",
     ],
 )
