+++ packages/zone.js/scripts/closure/closure_compiler.sh
@@ -1,7 +1,7 @@
 # compile closure test source file
 $(npm bin)/tsc -p .
 # Run the Google Closure compiler java runnable with zone externs
-java -jar ./node_modules/google-closure-compiler-java/compiler.jar --flagfile './scripts/closure/closure_flagfile' --externs './lib/closure/zone_externs.js' --externs './node_modules/@externs/nodejs/v8/global.js' --process_common_js_modules
+java -jar ./node_modules/google-closure-compiler-java/compiler.jar --flagfile './scripts/closure/closure_flagfile' --externs './build/zone_externs.js' --externs './node_modules/@externs/nodejs/v8/global.js' --process_common_js_modules
 
 # the names of Zone exposed API should be kept correctly with zone externs, test program should exit with 0.
 node build/closure/zone-closure-bundle.js
