+++ packages/zone.js/test/BUILD.bazel
@@ -1,7 +1,7 @@
-load("//tools:defaults.bzl", "jasmine_node_test", "ts_library")
+load("//tools:defaults.bzl", "jasmine_node_test", "ts_library", "zone_compatible_jasmine_node_test")
 load("//packages/zone.js/test:karma_test.bzl", "karma_test")
 
-package(default_visibility = ["//:__pkg__"])
+package(default_visibility = ["//:__subpackages__"])
 
 exports_files([
     "assets/sample.json",
@@ -67,14 +67,8 @@ ts_library(
 ts_library(
     name = "test_node_lib",
     testonly = True,
-    srcs = glob(["node/*.ts"]) + [
-        "node-env-setup.ts",
-        "test-env-setup-jasmine-no-patch-clock.ts",
-    ],
+    srcs = glob(["node/*.ts"]),
     deps = [
-        ":common_spec_env",
-        ":common_spec_srcs",
-        ":common_spec_util",
         "//packages/zone.js/lib",
         "@npm//@types/shelljs",
         "@npm//@types/systemjs",
@@ -87,28 +81,12 @@ ts_library(
 ts_library(
     name = "node_entry_point",
     testonly = True,
-    srcs = ["node_entry_point.init.ts"],
-    deps = [
-        ":common_spec_env",
-        ":common_spec_srcs",
-        ":common_spec_util",
-        "//packages/zone.js/lib",
-        "@npm//@types/shelljs",
-        "@npm//@types/systemjs",
-        "@npm//rxjs",
-        "@npm//shelljs",
-        "@npm//systemjs",
+    srcs = [
+        "node-env-setup.ts",
+        "node_entry_point.init.ts",
     ],
-)
-
-ts_library(
-    name = "node_entry_point_no_patch_clock",
-    testonly = True,
-    srcs = ["node_entry_point_no_patch_clock.init.ts"],
     deps = [
         ":common_spec_env",
-        ":common_spec_srcs",
-        ":common_spec_util",
         "//packages/zone.js/lib",
         "@npm//@types/shelljs",
         "@npm//@types/systemjs",
@@ -145,13 +123,12 @@ ts_library(
 ts_library(
     name = "node_error_disable_policy_entry_point",
     testonly = True,
-    srcs = ["node_error_disable_policy_entry_point.init.ts"],
+    srcs = [
+        "node_error_disable_policy.ts",
+        "node_error_disable_policy_entry_point.init.ts",
+    ],
     deps = [
-        ":common_spec_env",
-        ":common_spec_util",
-        ":error_spec_srcs",
         ":node_error_entry_point",
-        "//packages/zone.js/lib",
     ],
 )
 
@@ -161,8 +138,6 @@ ts_library(
     srcs = ["node_error_entry_point.init.ts"],
     deps = [
         ":common_spec_env",
-        ":common_spec_util",
-        ":error_spec_srcs",
         "//packages/zone.js/lib",
     ],
 )
@@ -170,33 +145,25 @@ ts_library(
 ts_library(
     name = "node_error_lazy_policy_entry_point",
     testonly = True,
-    srcs = ["node_error_lazy_policy_entry_point.init.ts"],
+    srcs = [
+        "node_error_lazy_policy.ts",
+        "node_error_lazy_policy_entry_point.init.ts",
+    ],
     deps = [
-        ":common_spec_env",
-        ":common_spec_util",
-        ":error_spec_srcs",
         ":node_error_entry_point",
-        "//packages/zone.js/lib",
     ],
 )
 
-jasmine_node_test(
+zone_compatible_jasmine_node_test(
     name = "test_node",
     bootstrap = [":node_entry_point"],
     deps = [
+        ":common_spec_srcs",
         ":test_node_lib",
     ],
 )
 
-jasmine_node_test(
-    name = "test_node_no_jasmine_clock",
-    bootstrap = [":node_entry_point_no_patch_clock"],
-    deps = [
-        ":test_node_lib",
-    ],
-)
-
-jasmine_node_test(
+zone_compatible_jasmine_node_test(
     name = "test_node_bluebird",
     bootstrap = [":node_bluebird_entry_point"],
     deps = [
@@ -207,11 +174,17 @@ jasmine_node_test(
 jasmine_node_test(
     name = "test_node_error_disable_policy",
     bootstrap = [":node_error_disable_policy_entry_point"],
+    deps = [
+        ":error_spec_srcs",
+    ],
 )
 
 jasmine_node_test(
     name = "test_node_error_lazy_policy",
     bootstrap = [":node_error_lazy_policy_entry_point"],
+    deps = [
+        ":error_spec_srcs",
+    ],
 )
 
 ts_library(
@@ -219,6 +192,7 @@ ts_library(
     testonly = True,
     srcs = ["npm_package/npm_package.spec.ts"],
     deps = [
+        "@npm//@bazel/runfiles",
         "@npm//@types/shelljs",
     ],
 )
