+++ packages/zone.js/test/common/Promise.spec.ts
@@ -63,31 +63,6 @@ describe(
         log = [];
       });
 
-      xit('should allow set es6 Promise after load ZoneAwarePromise', (done) => {
-        const ES6Promise = require('es6-promise').Promise;
-        const NativePromise = global[zoneSymbol('Promise')];
-
-        try {
-          global['Promise'] = ES6Promise;
-          Zone.assertZonePatched();
-          expect(global[zoneSymbol('Promise')]).toBe(ES6Promise);
-          const promise = Promise.resolve(0);
-          console.log('promise', promise);
-          promise
-              .then(value => {
-                expect(value).toBe(0);
-                done();
-              })
-              .catch(error => {
-                fail(error);
-              });
-        } finally {
-          global['Promise'] = NativePromise;
-          Zone.assertZonePatched();
-          expect(global[zoneSymbol('Promise')]).toBe(NativePromise);
-        }
-      });
-
       it('should pretend to be a native code', () => {
         expect(String(Promise).indexOf('[native code]') >= 0).toBe(true);
       });
