+++ packages/zone.js/test/node_bluebird_entry_point.init.ts
@@ -8,17 +8,14 @@
 
 // Must be loaded before zone loads, so that zone can detect WTF.
 import './test_fake_polyfill';
-
 // Setup tests for Zone without microtask support
 import '../lib/zone';
 import '../lib/common/promise';
 import '../lib/common/to-string';
 import '../lib/node/node';
 // Setup test environment
-require('@bazel/jasmine').boot();
 import './test-env-setup-jasmine';
 import './wtf_mock';
-
 import '../lib/zone-spec/async-test';
 import '../lib/zone-spec/fake-async-test';
 import '../lib/zone-spec/long-stack-trace';
@@ -27,7 +24,6 @@ import '../lib/zone-spec/sync-test';
 import '../lib/zone-spec/task-tracking';
 import '../lib/zone-spec/wtf';
 import '../lib/rxjs/rxjs';
-
 import '../lib/testing/promise-testing';
 
 const globalErrors = (jasmine as any).GlobalErrors;
