+++ packages/zone.js/test/node_entry_point.init.ts
@@ -16,18 +16,10 @@
 // Must be loaded before zone loads, so that zone can detect WTF.
 import './node-env-setup';
 import './test_fake_polyfill';
-
 // Setup tests for Zone without microtask support
 import '../lib/node/rollup-main';
-
-require('@bazel/jasmine').boot();
 // Zone symbol prefix is set to '__zone_symbol2__' in node-env-setup.ts.
 import './test-env-setup-jasmine';
-if (typeof global !== 'undefined' &&
-    (global as any)['__zone_symbol_test__fakeAsyncAutoFakeAsyncWhenClockPatched'] !== false) {
-  (global as any)['__zone_symbol_test__fakeAsyncAutoFakeAsyncWhenClockPatched'] = true;
-}
-
 import './wtf_mock';
 import '../lib/testing/zone-testing';
 import '../lib/zone-spec/task-tracking';
