+++ packages/zone.js/test/node_error_entry_point.init.ts
@@ -8,18 +8,11 @@
 
 // Must be loaded before zone loads, so that zone can detect WTF.
 import './test_fake_polyfill';
-
 // Setup tests for Zone without microtask support
 import '../lib/zone';
 import '../lib/common/promise';
 import '../lib/common/to-string';
-
-process.env['errorpolicy'] = (global as any)['__Zone_Error_ZoneJsInternalStackFrames_policy'] =
-    'disable';
-// Setup test environment
-require('@bazel/jasmine').boot();
 import './test-env-setup-jasmine';
-
 import './wtf_mock';
 import '../lib/common/error-rewrite';
 import '../lib/node/node';
@@ -31,5 +24,4 @@ import '../lib/zone-spec/sync-test';
 import '../lib/zone-spec/task-tracking';
 import '../lib/zone-spec/wtf';
 import '../lib/rxjs/rxjs';
-
 import '../lib/testing/promise-testing';
