+++ packages/zone.js/test/promise/promise-adapter.mjs
@@ -1,6 +1,7 @@
-require('../../build/lib/node/rollup-main');
+import '../../build/zone.umd.js';
+
 Zone[Zone.__symbol__('ignoreConsoleErrorUncaughtError')] = true;
-module.exports.deferred = function() {
+const deferred = function () {
   const p = {};
   p.promise = new Promise((resolve, reject) => {
     p.resolve = resolve;
@@ -9,10 +10,12 @@ module.exports.deferred = function() {
   return p;
 };
 
-module.exports.resolved = (val) => {
+const resolved = (val) => {
   return Promise.resolve(val);
 };
 
-module.exports.rejected = (reason) => {
+const rejected = (reason) => {
   return Promise.reject(reason);
 };
+
+export default {deferred, resolved, rejected};
