+++ packages/zone.js/test/zone-spec/clock-tests/BUILD.bazel
@@ -1,0 +1,58 @@
+load("//tools:defaults.bzl", "jasmine_node_test", "ts_library")
+
+ts_library(
+    name = "patched_init",
+    testonly = True,
+    srcs = [
+        "enable-clock-patch.ts",
+        "patched.init.ts",
+    ],
+    deps = [
+        "//packages/zone.js/test:node_entry_point",
+    ],
+)
+
+ts_library(
+    name = "test_patched_lib",
+    testonly = True,
+    srcs = [
+        "fake-async-patched-clock.spec.ts",
+    ],
+    deps = [
+        "//packages/zone.js/lib:zone_d_ts",
+    ],
+)
+
+ts_library(
+    name = "unpatched_init",
+    testonly = True,
+    srcs = [
+        "unpatched.init.ts",
+    ],
+    deps = [
+        "//packages/zone.js/test:node_entry_point",
+    ],
+)
+
+ts_library(
+    name = "test_unpatched_lib",
+    testonly = True,
+    srcs = [
+        "fake-async-unpatched-clock.spec.ts",
+    ],
+    deps = [
+        "//packages/zone.js/lib:zone_d_ts",
+    ],
+)
+
+jasmine_node_test(
+    name = "test_patched",
+    bootstrap = [":patched_init"],
+    deps = [":test_patched_lib"],
+)
+
+jasmine_node_test(
+    name = "test_unpatched",
+    bootstrap = [":unpatched_init"],
+    deps = [":test_unpatched_lib"],
+)
