+++ packages/zone.js/test/zone-spec/clock-tests/patched.init.ts
@@ -6,6 +6,5 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-// Note: Make sure to define the policy before loading the basic entry point.
-import './node_error_disable_policy';
-import './node_error_entry_point.init';
+import './enable-clock-patch';
+import '../../node_entry_point.init';
