+++ packages/zone.js/test/zone-spec/fake-async-test.spec.ts
@@ -25,13 +25,6 @@ function emptyRun() {
 
 (supportNode as any).message = 'support node';
 
-function supportClock() {
-  const _global: any = typeof window === 'undefined' ? global : window;
-  return typeof jasmine.clock === 'function' &&
-      _global[zoneSymbol('fakeAsyncAutoFakeAsyncWhenClockPatched')];
-}
-
-(supportClock as any).message = 'support patch clock';
 
 describe('FakeAsyncTestZoneSpec', () => {
   let FakeAsyncTestZoneSpec = (Zone as any)['FakeAsyncTestZoneSpec'];
@@ -1067,173 +1060,7 @@ describe('FakeAsyncTestZoneSpec', () => {
     });
   });
 
-  describe(
-      'fakeAsyncTest should work without patch jasmine.clock',
-      ifEnvSupports(
-          () => {
-            return !supportClock() && supportNode();
-          },
-          () => {
-            const fakeAsync = (Zone as any)[Zone.__symbol__('fakeAsyncTest')].fakeAsync;
-            let spy: any;
-            beforeEach(() => {
-              spy = jasmine.createSpy('timer');
-              jasmine.clock().install();
-            });
-
-            afterEach(() => {
-              jasmine.clock().uninstall();
-            });
-
-            it('should check date type correctly', fakeAsync(() => {
-                 const d: any = new Date();
-                 expect(d instanceof Date).toBe(true);
-               }));
-
-            it('should check date type correctly without fakeAsync', () => {
-              const d: any = new Date();
-              expect(d instanceof Date).toBe(true);
-            });
-
-            it('should tick correctly', fakeAsync(() => {
-                 jasmine.clock().mockDate();
-                 const start = Date.now();
-                 jasmine.clock().tick(100);
-                 const end = Date.now();
-                 expect(end - start).toBe(100);
-               }));
-
-            it('should tick correctly without fakeAsync', () => {
-              jasmine.clock().mockDate();
-              const start = Date.now();
-              jasmine.clock().tick(100);
-              const end = Date.now();
-              expect(end - start).toBe(100);
-            });
-
-            it('should mock date correctly', fakeAsync(() => {
-                 const baseTime = new Date(2013, 9, 23);
-                 jasmine.clock().mockDate(baseTime);
-                 const start = Date.now();
-                 expect(start).toBe(baseTime.getTime());
-                 jasmine.clock().tick(100);
-                 const end = Date.now();
-                 expect(end - start).toBe(100);
-                 expect(end).toBe(baseTime.getTime() + 100);
-                 expect(new Date().getFullYear()).toEqual(2013);
-               }));
-
-            it('should mock date correctly without fakeAsync', () => {
-              const baseTime = new Date(2013, 9, 23);
-              jasmine.clock().mockDate(baseTime);
-              const start = Date.now();
-              expect(start).toBe(baseTime.getTime());
-              jasmine.clock().tick(100);
-              const end = Date.now();
-              expect(end - start).toBe(100);
-              expect(end).toBe(baseTime.getTime() + 100);
-              expect(new Date().getFullYear()).toEqual(2013);
-            });
-
-            it('should handle new Date correctly', fakeAsync(() => {
-                 const baseTime = new Date(2013, 9, 23);
-                 jasmine.clock().mockDate(baseTime);
-                 const start = new Date();
-                 expect(start.getTime()).toBe(baseTime.getTime());
-                 jasmine.clock().tick(100);
-                 const end = new Date();
-                 expect(end.getTime() - start.getTime()).toBe(100);
-                 expect(end.getTime()).toBe(baseTime.getTime() + 100);
-               }));
-
-            it('should handle new Date correctly without fakeAsync', () => {
-              const baseTime = new Date(2013, 9, 23);
-              jasmine.clock().mockDate(baseTime);
-              const start = new Date();
-              expect(start.getTime()).toBe(baseTime.getTime());
-              jasmine.clock().tick(100);
-              const end = new Date();
-              expect(end.getTime() - start.getTime()).toBe(100);
-              expect(end.getTime()).toBe(baseTime.getTime() + 100);
-            });
-
-            it('should handle setTimeout correctly', fakeAsync(() => {
-                 setTimeout(spy, 100);
-                 expect(spy).not.toHaveBeenCalled();
-                 jasmine.clock().tick(100);
-                 expect(spy).toHaveBeenCalled();
-               }));
-
-            it('should handle setTimeout correctly without fakeAsync', () => {
-              setTimeout(spy, 100);
-              expect(spy).not.toHaveBeenCalled();
-              jasmine.clock().tick(100);
-              expect(spy).toHaveBeenCalled();
-            });
-          },
-          emptyRun));
-
-  describe('fakeAsyncTest should patch jasmine.clock', ifEnvSupports(supportClock, () => {
-             let spy: any;
-             beforeEach(() => {
-               spy = jasmine.createSpy('timer');
-               jasmine.clock().install();
-             });
-
-             afterEach(() => {
-               jasmine.clock().uninstall();
-             });
-
-             it('should check date type correctly', () => {
-               const d: any = new Date();
-               expect(d instanceof Date).toBe(true);
-             });
-
-             it('should get date diff correctly', () => {
-               const start = Date.now();
-               jasmine.clock().tick(100);
-               const end = Date.now();
-               expect(end - start).toBe(100);
-             });
-
-             it('should tick correctly', () => {
-               const start = Date.now();
-               jasmine.clock().tick(100);
-               const end = Date.now();
-               expect(end - start).toBe(100);
-             });
-
-             it('should mock date correctly', () => {
-               const baseTime = new Date(2013, 9, 23);
-               jasmine.clock().mockDate(baseTime);
-               const start = Date.now();
-               expect(start).toBe(baseTime.getTime());
-               jasmine.clock().tick(100);
-               const end = Date.now();
-               expect(end - start).toBe(100);
-               expect(end).toBe(baseTime.getTime() + 100);
-             });
-
-             it('should handle new Date correctly', () => {
-               const baseTime = new Date(2013, 9, 23);
-               jasmine.clock().mockDate(baseTime);
-               const start = new Date();
-               expect(start.getTime()).toBe(baseTime.getTime());
-               jasmine.clock().tick(100);
-               const end = new Date();
-               expect(end.getTime() - start.getTime()).toBe(100);
-               expect(end.getTime()).toBe(baseTime.getTime() + 100);
-             });
-
-             it('should handle setTimeout correctly', () => {
-               setTimeout(spy, 100);
-               expect(spy).not.toHaveBeenCalled();
-               jasmine.clock().tick(100);
-               expect(spy).toHaveBeenCalled();
-             });
-           }, emptyRun));
-
-  describe('fakeAsyncTest should patch rxjs scheduler', ifEnvSupports(supportClock, () => {
+  describe('fakeAsyncTest should patch rxjs scheduler', ifEnvSupports(() => isNode, () => {
              let FakeAsyncTestZoneSpec = (Zone as any)['FakeAsyncTestZoneSpec'];
              let testZoneSpec: any;
              let fakeAsyncTestZone: Zone;
