+++ packages/zone.js/tsconfig.json
@@ -1,7 +1,9 @@
 {
   "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2015",
+    "module": "esnext",
+    "target": "es2020",
+    "moduleResolution": "node",
+    "esModuleInterop": true,
     "outDir": "build",
     "inlineSourceMap": true,
     "inlineSources": true,
@@ -18,9 +20,7 @@
       "es2015.symbol",
       "es2015.symbol.wellknown"
     ],
-    "typeRoots": [
-      "node_modules/@types"
-    ],
+    "typeRoots": ["node_modules/@types"]
   },
   "exclude": [
     "node_modules",
